[add][fix] add: training_model.py, add function load_model(). fix: fix final checkpoint doesn't auto save issue, model path mark "final".

+++ src/training_model.py
@@ -1,7 +1,9 @@
+from torch import exp
 from torch.autograd.grad_mode import no_grad
 import os
 import time
 import torch
+from torch.optim import optimizer
 try:
     from src.Model.find_file_name import get_filenames
     from src.Model.Model_Perform_Tool import draw_plot
@@ -29,6 +31,8 @@ class HW3_Model(object):
         }
 
     def training(self, loader, val_loader=None, NUM_EPOCH=1, printPerformance=True, saveDir=None, checkpoint=0, bestModelSave=False):
+        epoch_start = len(self.performance_history['train_acc'])
+        NUM_EPOCH = epoch_start + NUM_EPOCH
         if saveDir is not None:
             self.saveDir = '{}/{}'.format(saveDir, time.strftime('%m%d-%H%M'))
             try:
@@ -37,9 +41,16 @@ class HW3_Model(object):
                 print("Fail to create the directory {} !".format(self.saveDir))
             else:
                 print("Successfully created the directory: {}".format(self.saveDir))
-            self.checkpoint = checkpoint
-
-        for epoch in range(1, NUM_EPOCH+1):
+        try:
+            if self.saveDir is not None:
+                if checkpoint > 0:
+                    self.checkpoint = checkpoint
+                else:
+                    self.checkpoint = NUM_EPOCH
+        except NameError:
+            self.saveDir = None
+            print("No save Dir path & checkpoint !")
+        for epoch in range(epoch_start+1, NUM_EPOCH+1):
             time_start = time.time()
             train_acc = 0.
             train_loss = 0.
@@ -66,47 +77,50 @@ class HW3_Model(object):
             self.performance_history['train_loss'].append(self.train_loss)
 
             # valiation test
+            info_word = ""
             if val_loader is not None:
                 self.valiating(val_loader)
-                if printPerformance:  # print valiation performance
-                    print('[{:>2d}/{}] {:.3f} sec, Train acc: {:.7f}, loss: {:.7f} | Val acc: {:.7f}, loss: {:.7f}'.format(
-                        epoch, NUM_EPOCH, time.time()-time_start,
-                        self.train_acc, self.train_loss, self.val_acc, self.val_loss))
-            elif printPerformance:  # print train performance
-                print('[{:>2d}/{}] {:.3f}sec, Train acc: {:.7f}, loss: {:.7f}'.format(
-                    epoch, NUM_EPOCH, time.time()-time_start,
-                    self.train_acc, self.train_loss))
-
-            # only valiating is work, self.best_model_epoch and bestModelSave has meaning
-            if val_loader is not None:
+                info_word = " | Val acc: {:.7f}, loss: {:.7f}".format(
+                    self.val_acc, self.val_loss)
                 if epoch == 1:
                     self.best_model_epoch = 1
 
-                if self.performance_history['val_loss'][self.best_model_epoch-1] >= self.val_loss:
-                    # save the best model
-                    if saveDir is not None and bestModelSave is True:
-                        filenames = get_filenames(
-                            self.saveDir, 'best_e{:03d}*.pickle'.format(self.best_model_epoch))
-                        [os.remove(filename) for filename in filenames]
-                        self.best_model_epoch = epoch
-                        path = '{}/best_e{:03d}_{}.pickle'.format(
-                            self.saveDir, epoch, str(self.val_loss)[:6])
-                        self.save_model(path, onlyParameters=True)
-                    else:
-                        self.best_model_epoch = epoch
+            # print train performance
+            if printPerformance:
+                print("[{:>2d}/{}] {:.3f}sec, Train acc: {:.7f}, loss: {:.7f}{}".format(
+                    epoch, NUM_EPOCH, time.time()-time_start,
+                    self.train_acc, self.train_loss, info_word))
+
+            # only valiating is work, self.best_model_epoch and bestModelSave has meaning
+            if val_loader is not None and self.performance_history['val_loss'][self.best_model_epoch-1] >= self.val_loss:
+                if bestModelSave is True:
+                    filenames = get_filenames(
+                        self.saveDir, 'best_e{:03d}*.pickle'.format(self.best_model_epoch))
+                    [os.remove(filename) for filename in filenames]
+                    self.best_model_epoch = epoch
+                    path = '{}/best_e{:03d}_{}.pickle'.format(
+                        self.saveDir, epoch, str(self.val_loss)[:6])
+                    self.save_model(path, onlyParameters=True)
+                else:
+                    self.best_model_epoch = epoch
 
             # checkpoint save model
-            if epoch % self.checkpoint == 0:
-                path = ''
-                # path name rule is: mmdd-hhmm_epoch_loss-value, e.g. 0814-1600_e010_0.00138.pickle
+            mark_word = ''
+            MODEL_SAVE = True
+            if epoch == NUM_EPOCH:
+                mark_word = 'final_'
+            elif epoch % self.checkpoint == 0:
+                mark_word = ''
+            else:
+                MODEL_SAVE = False
+            # path name rule is: (mark_name)_e(epoch)_(loss-value).pickle, e.g. best_e010_0.00138.pickle
+            if MODEL_SAVE is True and self.saveDir is not None:
                 if val_loader is not None:
-                    # loss-value = val_loss
-                    path = '{}/e{:03d}_{}.pickle'.format(
-                        self.saveDir, epoch, str(self.val_loss)[:6])
+                    LOSS_VALUE = self.val_loss
                 else:
-                    # loss-value = train_loss
-                    path = '{}/e{:03d}_{}.pickle'.format(
-                        self.saveDir, epoch, str(self.train_loss)[:6])
+                    LOSS_VALUE = self.train_loss
+                path = '{}/{}e{:03d}_{}.pickle'.format(
+                    self.saveDir, mark_word, epoch, str(LOSS_VALUE)[:6])
                 self.save_model(path, onlyParameters=True)
 
     def valiating(self, loader):
@@ -149,17 +163,32 @@ class HW3_Model(object):
         if onlyParameters:
             # torch.save(self.net.state_dict(), path)
             net = self.net
+            optimizer = self.optimizer
             self.net = self.net.state_dict()
+            self.optimizer = self.optimizer.state_dict()
             # with open(path, 'wb') as target:
             #     pickle.dump(self, target)
             # self.net = net
             torch.save(self, path)
             self.net = net
+            self.optimizer = optimizer
 
         else:
             # with open(path, 'wb') as target:
             #     pickle.dump(self, path)
             torch.save(self, path)
+
+    def load_model(self, path):
+        model = torch.load(path)
+        try:
+            self.best_model_epoch = model.best_model_epoch
+        except:
+            pass
+        self.saveDir = model.saveDir
+        self.performance_history = model.performance_history
+        self.net.load_state_dict(model.net)
+        self.optimizer.load_state_dict(model.optimizer)
+        self.net.eval()
 
     def get_performance_plt(self):
         draw_plot(self.performance_history['train_acc'], self.performance_history['train_loss'],
